requirements.txt
Click == 7.0
albumentations == 0.2.3
connected_components_3d == 1.4.1
matplotlib == 3.0.3
nibabel == 2.4.0
numpy == 1.16.3
opencv_python_headless == 4.1.0.25
pathlib2 == 2.3.3
scipy == 1.1.0
tensorboardX == 1.6
torchsummary == 1.5.1



torchvision